- PR Feedback, Joel   - create Time/Utils.tsx   - fix (maybe) duration dropdown/RateIntervals.getName() - Increased the play/pause button size - got the datetimepicker working and simplified times by trimming away the seconds. - changed the graph border to solid, made it just the left side, and extended it to group the replay toolbar - migrated the "duration" dropdown back to the top toolbar and changed the text from, for example "Last 5m" to "5m Traffic". Duration is not a replay-specific setting and should probably stay in the top toolbar. I changed the text to hopefully be applicable to both normal and replay scenarios. Also, "Traffic" is more descriptive of what the setting actually does; it sets the amount of traffic data pulled per rendering. The tooltips are situational. - Pulled the "floating X" up to the top toolbar and gave it a text description. Now the open and close of replay is located on the same horizontal. - as a side-effect of the changes the progress bar gets more space. - updated broken snapshots
